* replaced QString conversions using toUtf8() with toLocale8Bit() hoping to support more platforms and variants * removed warnings; compiler, don't cry!
git-svn-id: http://svn.code.sf.net/p/qsynth/code_attic/trunk@289 6e21f190-4f5b-40a6-a6af-ed3d11dd9ba7

+++ src/qsynthDialPeppinoStyle.cpp
@@ -1,11 +1,11 @@
 /***************************************************************************
 
-   This style is based on a widget designed by Giuseppe Cigala,  
+   This style is based on a widget designed by Giuseppe Cigala,
    adapted as a Qt style for QSynth by Pedro Lopez-Cabanillas
 
    This file,
-   Copyright (C) 2008 Giuseppe Cigala <g_cigala@virgilio.it>,          
-   Copyright (C) 2008 Pedro Lopez-Cabanillas <plcl@users.sf.net>         
+   Copyright (C) 2008 Giuseppe Cigala <g_cigala@virgilio.it>,
+   Copyright (C) 2008 Pedro Lopez-Cabanillas <plcl@users.sf.net>
 
    This program is free software; you can redistribute it and/or
    modify it under the terms of the GNU General Public License
@@ -72,13 +72,13 @@ inline void paintBorder(QPainter *p)
 
 inline int valueAngle(const QStyleOptionSlider *dial)
 {
-    return -((dial->sliderValue - dial->minimum) * 4320) / (dial->maximum - dial->minimum);	
+    return -((dial->sliderValue - dial->minimum) * 4320) / (dial->maximum - dial->minimum);
 }
 
 inline void paintArc(QPainter *p, const QStyleOptionSlider *dial)
 {
     QPalette pal = dial->palette;
-    QColor arcColor = (dial->state & QStyle::State_Enabled) ? pal.highlight().color() : pal.mid().color();       
+    QColor arcColor = (dial->state & QStyle::State_Enabled) ? pal.highlight().color() : pal.mid().color();
     QLinearGradient linGrad(80, 100, 140, 140);
     linGrad.setColorAt(0, arcColor.dark(140));
     linGrad.setColorAt(1, arcColor.light().light());
@@ -87,7 +87,7 @@ inline void paintArc(QPainter *p, const QStyleOptionSlider *dial)
 
     p->setPen(QPen(Qt::black, 1, Qt::SolidLine, Qt::RoundCap, Qt::RoundJoin));
     QRectF rectangle(15, 15, 170, 170);
-    
+
     int spanAngle = valueAngle(dial);
     int startAngle = 225 * 16;
     p->drawPie(rectangle, startAngle, spanAngle);
@@ -115,12 +115,12 @@ inline void paintDot(QPainter *p, const QStyleOptionSlider *dial)
 {
     int startPoint = (225 * 16) + valueAngle(dial);
     QPalette pal = dial->palette;
-    QPen dotPen((dial->state & QStyle::State_Enabled) ? Qt::red : pal.mid().color(), 7, Qt::SolidLine, Qt::RoundCap, Qt::RoundJoin); 
+    QPen dotPen((dial->state & QStyle::State_Enabled) ? Qt::red : pal.mid().color(), 7, Qt::SolidLine, Qt::RoundCap, Qt::RoundJoin);
     QRect rectangle1(45, 45, 110, 110);
     p->setPen(dotPen);
     p->drawArc(rectangle1, startPoint, 5);
 
-    if (dial->subControls & QStyle::SC_DialTickmarks) 
+    if (dial->subControls & QStyle::SC_DialTickmarks)
     {
     	QRect rectangle2(10, 10, 180, 180);
     	int ns = dial->tickInterval;
@@ -129,34 +129,34 @@ inline void paintDot(QPainter *p, const QStyleOptionSlider *dial)
     	p->setPen(QPen(Qt::black, 5, Qt::SolidLine, Qt::RoundCap, Qt::RoundJoin));
     	for (int i = 0; i <= dot; i++)
     	{
-    		p->drawArc(rectangle2, 225*16 - delta*i, 5);
+    		p->drawArc(rectangle2, int(225*16 - delta*i), 5);
     	}
     }
 }
 
-void 
+void
 qsynthDialPeppinoStyle::drawComplexControl(ComplexControl cc, const QStyleOptionComplex *opt, QPainter *p, const QWidget *widget) const
 {
 	if (cc != QStyle::CC_Dial) {
 		QCommonStyle::drawComplexControl(cc, opt, p, widget);
 		return;
 	}
-	
-	const QStyleOptionSlider *dial = qstyleoption_cast<const QStyleOptionSlider *>(opt); 
+
+	const QStyleOptionSlider *dial = qstyleoption_cast<const QStyleOptionSlider *>(opt);
 	if (dial == NULL)
 		return;
-	
+
 	p->save();
 	int size = dial->rect.width() < dial->rect.height() ? dial->rect.width() : dial->rect.height();
 	p->setViewport((dial->rect.width()-size)/2, (dial->rect.height()-size)/2, size, size);
     p->setWindow(0, 0, 200, 200);
     p->setRenderHint(QPainter::Antialiasing);
-	
+
     paintBorder(p);
     paintArc(p, dial);
     paintDial(p);
     paintDot(p, dial);
-	
+
 	// done
-	p->restore();	
+	p->restore();
 }
